feat(gpio): GPIO source file and init function
Create the header file of the GPIO driver with the tentative function
declarations.
Define the structs for configure the GPIOx registers and the pin modes,
besides provide some typedef enumerations that will hold the options
availale for configuring the pin.
Define a Driver status enumerations to indicate if the functions returns
succesfully or there is an error.

JIRA: SCRUM-5, SCRUM-7