feat(gpio): GPIO function implementations
Implement the following functions to the gpio driver source file:

  * Initialization
  * Write to pin
  * Read from pin
  * Clock control

In order to use GPIO peripheral driver, in the header file write typedefs
to select between distinct modes and configuration. To hold
the information of the configuration and the instance of the GPIO,
define GPIO_ConfigTypeDef  and GPIO_PinConfig to configure the pin as
desired.

JIRA: SCRUM-5, SCRUM-9, SCRUM-12