feat(display): Add initialization function
Add a function that will follow the initialization sequence suggested in
the datasheet of the display. Include two functions that will fill the
screen entirely with black pixels and white pixels.
Add a SysTick initialization function that will enable the SystTick
interruption in order to keep track of the ellapsed time. Implement with
his a delay function that will halt the processor during a desired
delay.

JIRA: Scrum-21