feat(display): Create image data
Create an array of size 4000 bytes (width / 8 * height) that will represent the image displayed on the e-ink
paper. This array will be modified in order to achieve different images.
Implement a function that will display the current image stored. In
order to test this, show in the display two lines by modifying this
array.
Corrected minor bugs that initialazed improperly the display.

JIRA: SCRUM-22